% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start.R
\name{collapse_named_lists}
\alias{collapse_named_lists}
\title{Collapse a list of named lists into one list with the same name}
\usage{
collapse_named_lists(in_list)
}
\arguments{
\item{in_list}{a list of named lists all with same names (not checked for efficiency)}
}
\value{
a single list with named components all concatenated
}
\description{
Collapse a list of named lists into one list with the same name
}
