% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_row}
\alias{make_row}
\title{Build a block row matrix for multiview}
\usage{
make_row(x_list, p_x, pair, rho)
}
\arguments{
\item{x_list}{list of x matrices}

\item{p_x}{a list of ncol of elements in x_list}

\item{pair}{an integer vector of two indices}

\item{rho}{the rho value}
}
\value{
a block row of matrix for multiview
}
\description{
Build a block row matrix for multiview
}
