% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiview-package.R
\name{multiview-package}
\alias{multiview-package}
\title{Cooperative learning for multiple views using generalized linear models}
\description{
This package performs a version of early and late fusion of multiple views using penalized generalized regression.
}
\keyword{models}
\keyword{package}
\keyword{regression}
