% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiview-package.r
\docType{package}
\name{multiview}
\alias{multiview}
\title{multiview: Methods for High-Dimensional Multi-View Learning.}
\description{
The multiview package currently provides a function StaPLR which fits a stacked penalized logistic regression model.
}
\author{
Wouter van Loon <w.s.van.loon@fsw.leidenuniv.nl>
}
