% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmds.R
\name{mvmds}
\alias{mvmds}
\title{Multiview MDS on a list of matrices or \code{dist} objects.}
\usage{
mvmds(x, k = 2)
}
\arguments{
\item{x}{A list of data matrices or dist objects. Both types can be mixed.
In the case of plain data matrices, euclidean distance will be used to
generate a \code{dist} object for that data view.}

\item{k}{Number of desired dimensions of the low-dimensional projection.}
}
\value{
A \code{n x k} matrix with the k-dimensional projection, where \code{n}
   is the number of samples in the dataset.
}
\description{
Multiview multidimensional scaling (\code{mvmds}) receives two or more
feature matrices or \code{dist} objects
(or any combination of both) and produces a low-dimensional representation
of the samples according to the combined information in all the input data.
}
\note{
All input views must have the same number of samples (rows).
}
\examples{
x1 <- iris[, 1:2]
x2 <- iris[, 3:4]
mvmds(list(x1, x2), k = 2)

}
