% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsc.R
\name{mvsc}
\alias{mvsc}
\title{Multiview spectral clustering on a list of matrices or \code{dist} objects.}
\usage{
mvsc(x, k, sigmas = NULL, neighbours = NULL, clustering = TRUE)
}
\arguments{
\item{x}{A list of feature matrices or \code{dist} objects (or a mix of both).}

\item{k}{Number of desired clusters.}

\item{sigmas}{Either \code{NULL}, a single real value or a vector of real values.
They correspond
to the \code{sigma} parameter in the Gaussian radial basis function.
If it is NULL then the default sigma computation is used (average
distance to the log(n)-th neighbour, with n = number of samples), unless
\code{neighbours} has a value different from \code{NULL}.
If it is a single number then the same sigma is applied to all input
views. If it is a vector each value in it is applied to the corresponding
input view.}

\item{neighbours}{Either \code{NULL}, a single integer value or a vector of integers.
They correspond
to the expected number of neighbours per point, used to estimate the
\code{sigma} values of the Gaussian radial basis function.
If it is NULL then the default sigma computation is used (average
distance to the log(n)-th neighbour, with n = number of samples).
If it is a single value then the same number of neighbours is used
on all input views, else each value in the vector is applied to the
corresponding input view.
Does not have effect if \code{sigma} is different from \code{NULL}.}

\item{clustering}{Tells \code{mvsc} if it has to perform the clustering on the projection
(if TRUE) or to skip the clustering step of the algorithm.}
}
\value{
A list with four elements: \code{clustering} is a vector of integers with the
  clustering assignment of each sample (not included if \code{clustering = FALSE}),
  \code{evalues} is a matrix with the eigenvalues
  of the common principal components (CPC) step, \code{evectors} is a matrix with the
  eigenvectors of the CPC step, and \code{sigmas} is a vector with the sigmas used on
  the Gaussian radial basis function of each input view.
}
\description{
Computes the multiview spectral clustering of data on a list of matrices
or dist objects (or a mix of both), supposed to be different views of the same data.
}
\note{
All input views must have the same number of samples (rows).
}
\examples{
m1 <- iris[, 1:2]
m2 <- iris[, 3:4]
mvsc(list(m1, m2), k = 3)

}
