% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StaPLR.R
\name{predict.StaPLR}
\alias{predict.StaPLR}
\title{Make predictions from a "StaPLR" object.}
\usage{
\method{predict}{StaPLR}(object, newx, metadat = "response",
  predtype = "response", cvlambda = "lambda.min")
}
\arguments{
\item{object}{Fitted "StaPLR" model object.}

\item{newx}{Matrix of new values for x at which predictions are to be made. Must be a matrix.}

\item{metadat}{The attribute of the base-learners to be used as input to the meta-learner.}

\item{predtype}{The type of prediction returned by the meta-learner.}

\item{cvlambda}{Value of the penalty parameter lambda at which predictions are to be made.}
}
\value{
TBA.
}
\description{
Fit a two-level stacked penalized logistic regression model with a single base-learner and a single meta-learner.
}
\examples{
TBA
}
\author{
Wouter van Loon <w.s.van.loon@fsw.leidenuniv.nl>
}
\keyword{TBA}
