% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start.R
\name{reshape_x_to_xlist}
\alias{reshape_x_to_xlist}
\title{Return a new list of x matrices of same shapes as those in x_list}
\usage{
reshape_x_to_xlist(x, x_list)
}
\arguments{
\item{x}{the column-binded entries of \code{x_list}}

\item{x_list}{a list of \code{x} matrices with same number of rows
\code{nobs}}
}
\description{
Return a new list of x matrices of same shapes as those in x_list
}
