% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start.R
\name{select_matrix_list_columns}
\alias{select_matrix_list_columns}
\title{Select x_list columns specified by (conformable) list of indices}
\usage{
select_matrix_list_columns(x_list, indices)
}
\arguments{
\item{x_list}{a list of \code{x} matrices with same number of rows
\code{nobs}}

\item{indices}{a vector of indices in \code{1:nvars}}
}
\value{
a list of x matrices
}
\description{
Select x_list columns specified by (conformable) list of indices
}
