% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_nvar_index}
\alias{to_nvar_index}
\title{Translate from column indices in list of x matrices to indices in
\code{1:nvars}. No sanity checks for efficiency}
\usage{
to_nvar_index(x_list, index_list)
}
\arguments{
\item{x_list}{a list of \code{x} matrices with same number of rows
\code{nobs}}

\item{index_list}{a list of column indices for each matrix,
including possibly column indices of length 0}
}
\value{
a vector of indices between 1 and nvars = sum of \code{ncol(x)}
for x in x_list
}
\description{
Translate from column indices in list of x matrices to indices in
\code{1:nvars}. No sanity checks for efficiency
}
