% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start.R
\name{to_xlist_index}
\alias{to_xlist_index}
\title{Translate indices in \code{1:nvars} to column indices in list of x
matrices. No sanity checks}
\usage{
to_xlist_index(x_list, index)
}
\arguments{
\item{x_list}{a list of \code{x} matrices with same number of rows
\code{nobs}}

\item{index}{vector of indices between 1 and nvars = sum of
\code{ncol(x)} for x in x_list}
}
\value{
a conformed list of column indices for each matrix,
including possibly column indices of length 0
}
\description{
Translate indices in \code{1:nvars} to column indices in list of x
matrices. No sanity checks
}
